requirements.txt
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
distro==1.7.0
Django==4.0.3
django-filter==21.1
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==3.2.0
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.10.3
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
pytest==7.1.1
pytest-django==4.5.2
PyYAML==5.4.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.9
websocket-client==0.59.0
